+++ src/note_widget/CodeBlock.js
@@ -1,14 +1,22 @@
 import React from 'react'
 import { string } from 'prop-types'
-import hljs from 'highlight.js/lib/highlight';
+import hljs from 'highlight.js/lib/core';
 import 'highlight.js/styles/github.css';
 import javascript from 'highlight.js/lib/languages/javascript';
 import java from 'highlight.js/lib/languages/java';
 import python from 'highlight.js/lib/languages/python';
+import css from 'highlight.js/lib/languages/css';
+import html from 'highlight.js/lib/languages/htmlbars';
+import sql from 'highlight.js/lib/languages/sql';
+import cpp from 'highlight.js/lib/languages/cpp';
+
 hljs.registerLanguage('javascript', javascript);
+hljs.registerLanguage('html', html);
 hljs.registerLanguage('java', java);
 hljs.registerLanguage('python', python);
-
+hljs.registerLanguage('css', css);
+hljs.registerLanguage('sql', sql);
+hljs.registerLanguage('cpp', cpp);
 
 
 // Code adopted from https://github.com/rexxars/react-markdown/blob/master/demo/src/code-block.js