+++ src/options_page/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 // import logo from './logo.svg';
 import './App.css';
-import Nav from './options_page/Nav.js';
+import Nav from './Nav.js';
 
 
 class App extends Component {